% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{AIC}
\alias{AIC}
\alias{AIC.BranchGLM}
\alias{BIC.BranchGLM}
\title{Extract AIC}
\usage{
\method{AIC}{BranchGLM}(fit)

\method{BIC}{BranchGLM}(fit)
}
\arguments{
\item{fit}{A BranchGLM or BranchGLMboot object.}
}
\value{
Returns AIC
}
\description{
Gets AIC from fitted model.
}
\examples{
Data <- iris
Fit <- BranchGLM(Sepal.Length ~ ., data = Data, family = "gaussian", link = "identity")
AIC(Fit)
BIC(Fit)
}
