% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{BIC.BranchGLM}
\alias{BIC.BranchGLM}
\title{Extract BIC}
\usage{
\method{BIC}{BranchGLM}(object, ...)
}
\arguments{
\item{object}{a \code{BranchGLM} model object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract BIC
}
