% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM-package.R
\name{BranchGLM-defunct}
\alias{BranchGLM-defunct}
\alias{fit.summary.BranchGLMVS}
\alias{fit.BranchGLMVS}
\title{Defunct functions in BranchGLM}
\usage{
fit.summary.BranchGLMVS(...)

fit.BranchGLMVS(...)
}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are: \link{fit.BranchGLMVS} and \link{fit.summary.BranchGLMVS}
}
\keyword{internal}
