% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM-package.R
\name{BranchGLM-internal}
\alias{BranchGLM-internal}
\alias{BranchGLMFit}
\alias{MetricIntervalCpp}
\alias{SwitchBranchAndBoundCpp}
\alias{BranchAndBoundCpp}
\alias{BackwardBranchAndBoundCpp}
\alias{ForwardCpp}
\alias{BackwardCpp}
\alias{FastBackwardCpp}
\alias{DoubleBackwardCpp}
\alias{FastDoubleBackwardCpp}
\alias{MakeTable}
\alias{MakeTableFactor2}
\alias{CindexCpp}
\alias{CindexTrap}
\alias{ROCCpp}
\alias{SwitchVariableImportanceCpp}
\title{Internal BranchGLM Functions}
\description{
Internal BranchGLM Functions.
}
\details{
These are not intended for use by users, these are Rcpp functions
that do not check the arguments, so improper usage may result in R crashing.
}
\keyword{internal}
