% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{BranchGLM}
\alias{BranchGLM}
\title{Fits GLMs}
\usage{
BranchGLM(
  formula,
  data,
  family,
  link,
  offset = NULL,
  method = "Fisher",
  grads = 10,
  parallel = FALSE,
  nthreads = 8,
  tol = 1e-04,
  contrasts = NULL
)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A dataframe.}

\item{family}{distribution used to model the data, one of "gaussian", "binomal", or "poisson"}

\item{link}{link used to link mean structure to linear predictors. One of, 
"identity", "logit", "probit", "cloglog", or "log".}

\item{offset}{Offset vector, by default the zero vector is used.}

\item{method}{One of "Fisher", "BFGS", or "LBFGS".}

\item{grads}{Number of gradients to use to approximate information with, only for LBFGS.}

\item{parallel}{Whether or not to make use of parallelization via OpenMP.}

\item{nthreads}{Number of threads used with OpenMP, only used if parallel = TRUE.}

\item{tol}{Tolerance used to determine model convergence.}

\item{contrasts}{see contrasts.arg of model.matrix.default.}
}
\value{
A BranchGLM object.
}
\description{
Fits a GLM via RcppArmadillo, can be fit in parallel.
}
\details{
The likelihood equations are solved directly when using any of the methods 
i.e. no matrix decomposition is used.

Can use BFGS, L-BFGS, or Fisher Scoring to fit the GLM. BFGS and LBFGS are 
typically faster than Fisher Scoring when there are at least 50 covariates 
and Fisher Scoring is typically best when there are fewer than 20 covariates.
}
