% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{coef}
\alias{coef}
\alias{coef.BranchGLM}
\title{Extract Coefficient Estimates}
\usage{
\method{coef}{BranchGLM}(fit)
}
\arguments{
\item{fit}{A BranchGLM or BranchGLMboot object.}
}
\value{
Returns a named vector of the coefficient estimates.
}
\description{
Extract coefficient estimates from BranchGLM or BranchGLMboot object.
}
\examples{
Data <- iris
Fit <- BranchGLM(Sepal.Length ~ ., data = Data, family = "gaussian", link = "identity")
coef(Fit)
}
