% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{deviance.BranchGLM}
\alias{deviance.BranchGLM}
\title{Extract the Deviance}
\usage{
\method{deviance}{BranchGLM}(object, ...)
}
\arguments{
\item{object}{a \code{BranchGLM} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A single number denoting the deviance.
}
\description{
Extracts the deviance from \code{BranchGLM} objects.
}
