% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{family.BranchGLM}
\alias{family.BranchGLM}
\title{Extract Family from BranchGLM Objects}
\usage{
\method{family}{BranchGLM}(object, ...)
}
\arguments{
\item{object}{a \code{BranchGLM} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \code{family}.
}
\description{
Extracts family from \code{BranchGLM} objects.
}
\note{
Only the \code{family} and \code{link} components of \code{family} objects are used by the
\code{BranchGLM} package. The other components of the \code{family} object may not
be the same as what is used in the fitting process for \code{BranchGLM}.
}
