% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection.R, R/summaryBranchGLMVS.R
\name{fit.BranchGLMVS}
\alias{fit.BranchGLMVS}
\alias{fit}
\alias{fit.summary.BranchGLMVS}
\title{Fits GLMs for summary.BranchGLMVS and BranchGLMVS Objects}
\usage{
\method{fit}{BranchGLMVS}(object, which = 1, keepData = TRUE, keepY = TRUE, ...)

fit(object, ...)

\method{fit}{summary.BranchGLMVS}(object, which = 1, keepData = TRUE, keepY = TRUE, ...)
}
\arguments{
\item{object}{a \code{summary.BranchGLMVS} or \code{BranchGLMVS} object.}

\item{which}{a positive integer indicating which model to fit,
the default is to fit the first model .}

\item{keepData}{Whether or not to store a copy of data and design matrix, the default
is TRUE. If this is FALSE, then the results from this cannot be used inside of \code{VariableSelection}.}

\item{keepY}{Whether or not to store a copy of y, the default is TRUE. If
this is FALSE, then the binomial GLM helper functions may not work and this
cannot be used inside of \code{VariableSelection}.}

\item{...}{further arguments passed to other methods.}
}
\value{
An object of class \link{BranchGLM}.
}
\description{
Fits GLMs for summary.BranchGLMVS and BranchGLMVS Objects
}
\details{
The information needed to fit the GLM is taken from the original information
supplied to the \code{VariableSelection} function.

The fitted models do not have standard errors or p-values since these are
biased due to the selection process.
}
