% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{logLik.BranchGLM}
\alias{logLik.BranchGLM}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{BranchGLM}(object, ...)
}
\arguments{
\item{object}{a \code{BranchGLM} model object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract Log-Likelihood
}
