% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{logLik}
\alias{logLik}
\alias{logLik.BranchGLM}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{BranchGLM}(fit)
}
\arguments{
\item{fit}{A BranchGLM or BranchGLMboot object.}
}
\value{
Returns log-likelihood
}
\description{
Gets log-likelihood from fitted model.
}
