% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{model.frame.BranchGLM}
\alias{model.frame.BranchGLM}
\title{Extract Model Frame from a BranchGLM Object}
\usage{
\method{model.frame}{BranchGLM}(formula, ...)
}
\arguments{
\item{formula}{a \code{BranchGLM} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \link{data.frame} used for the fitted model.
}
\description{
Extracts model frame from \code{BranchGLM} objects.
}
