% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{nobs.BranchGLM}
\alias{nobs.BranchGLM}
\title{Extract Number of Observations from BranchGLM Objects}
\usage{
\method{nobs}{BranchGLM}(object, ...)
}
\arguments{
\item{object}{a \code{BranchGLM} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A single number indicating the number of observations used to fit the model.
}
\description{
Extracts number of observations from BranchGLM objects.
}
