% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{plot.BranchGLM}
\alias{plot.BranchGLM}
\title{Plot Method for BranchGLM Objects}
\usage{
\method{plot}{BranchGLM}(x, ...)
}
\arguments{
\item{x}{a \code{BranchGLM} object.}

\item{...}{further arguments passed to \link{plot.default}.}
}
\value{
This only produce a plot, nothing is returned.
}
\description{
Creates a plot to help visualize fitted values from BranchGLM objects.
}
\examples{
Data <- iris
Fit <- BranchGLM(Sepal.Length ~ ., data = Data, family = "gaussian", link = "identity")
plot(Fit)

}
