% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{predict}
\alias{predict}
\alias{predict.BranchGLM}
\title{Predict Method for BranchGLM Object}
\usage{
\method{predict}{BranchGLM}(fit, newdata = NULL, type = "response")
}
\arguments{
\item{fit}{A BranchGLM or BranchGLMboot object.}

\item{newdata}{A dataframe, if not specified the data the model was fit on is used.}

\item{type}{One of "linpreds" or "response", if not specified "response" is used.}

\item{contrasts}{see contrasts.arg of model.matrix.default.}
}
\value{
A numeric vector of predictions.
}
\description{
Gets predictions from BranchGLM model.
}
\details{
Link corresponds to the linear predictors and response corresponds to the predicted probabilities.
}
