% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{print.BranchGLM}
\alias{print.BranchGLM}
\title{Print Method for BranchGLM}
\usage{
\method{print}{BranchGLM}(x, coefdigits = 4, digits = 0, ...)
}
\arguments{
\item{x}{a \code{BranchGLM} model object.}

\item{coefdigits}{number of digits to display for coefficients table.}

\item{digits}{number of digits to display for information after table.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print Method for BranchGLM
}
