% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMTable.R
\name{print.BranchGLMTable}
\alias{print.BranchGLMTable}
\title{Print Method for BranchGLMTable}
\usage{
\method{print}{BranchGLMTable}(Table, digits = 4)
}
\arguments{
\item{Table}{A BranchGLMTable object.}

\item{digits}{Number of digits to display.}
}
\description{
Print Method for BranchGLMTable
}
