% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableImportance.boot.R
\name{print.BranchGLMVI.boot}
\alias{print.BranchGLMVI.boot}
\title{Print Method for BranchGLMVI.boot Objects}
\usage{
\method{print}{BranchGLMVI.boot}(x, digits = 3, ...)
}
\arguments{
\item{x}{a \code{BranchGLMVI.boot} object.}

\item{digits}{number of significant digits to display.}

\item{...}{further arguments passed to other methods.}
}
\value{
The supplied \code{BranchGLMVI.boot} object.
}
\description{
Print method for BranchGLMVI.boot objects.
}
