% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableImportance.R
\name{print.BranchGLMVI}
\alias{print.BranchGLMVI}
\title{Print Method for BranchGLMVI Objects}
\usage{
\method{print}{BranchGLMVI}(x, digits = 3, ...)
}
\arguments{
\item{x}{a \code{BranchGLMVI} object.}

\item{digits}{number of significant digits to display.}

\item{...}{further arguments passed to other methods.}
}
\value{
The supplied \code{BranchGLMVI} object.
}
\description{
Print method for BranchGLMVI objects.
}
