% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection.R
\name{print.BranchGLMVS}
\alias{print.BranchGLMVS}
\title{Print Method for BranchGLMVS}
\usage{
\method{print}{BranchGLMVS}(VS, coefdigits = 4, digits = 0)
}
\arguments{
\item{coefdigits}{Number of digits to display for coefficients table.}

\item{digits}{Number of digits to display for information after table.}

\item{fit}{A BranchGLMVS model object.}
}
\description{
Print Method for BranchGLMVS
}
