% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBranchGLMVS.R
\name{print.summary.BranchGLMVS}
\alias{print.summary.BranchGLMVS}
\title{Print Method for summary.BranchGLMVS}
\usage{
\method{print}{summary.BranchGLMVS}(x, digits = 4, ...)
}
\arguments{
\item{x}{a \code{summary.BranchGLMVS} object.}

\item{digits}{number of digits to display for information in the table.}

\item{...}{further arguments passed to other methods.}
}
\value{
The supplied \code{summary.BranchGLMVS} object.
}
\description{
Print Method for summary.BranchGLMVS
}
