% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{residuals.BranchGLM}
\alias{residuals.BranchGLM}
\title{Extract the Pearson Residuals from BranchGLM Objects}
\usage{
\method{residuals}{BranchGLM}(object, ...)
}
\arguments{
\item{object}{a \code{BranchGLM} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A vector of the Pearson residuals from the supplied \code{BranchGLM} object.
}
\description{
Extracts the Pearson residuals from \code{BranchGLM} objects.
}
