% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMTable.R
\name{ROC}
\alias{ROC}
\alias{ROC.numeric}
\alias{ROC.BranchGLM}
\title{ROC Curve}
\usage{
ROC(x, ...)

\method{ROC}{numeric}(preds, y)

\method{ROC}{BranchGLM}(fit)
}
\arguments{
\item{preds}{A numeric vector of predictions between 0 and 1.}

\item{y}{Observed values, can be a numeric vector of 0s and 1s, a two-level 
factor vector, or a logical vector.}

\item{fit}{A \code{BranchGLM} object.}
}
\value{
A \code{BranchGLMROC} object which can be plotted with \code{plot()}. The AUC can also 
be calculated using \code{AUC()}.
}
\description{
Creates an ROC curve.
}
\examples{
Data <- ToothGrowth
Fit <- BranchGLM(supp ~ ., data = Data, family = "binomial", link = "logit")
MyROC <- ROC(Fit)
plot(MyROC)
}
