% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection.R
\name{VariableSelection}
\alias{VariableSelection}
\alias{VariableSelection.formula}
\alias{VariableSelection.BranchGLM}
\title{Variable Selection for GLMs}
\usage{
VariableSelection(x, ...)

\method{VariableSelection}{formula}(
  formula,
  data,
  family,
  link,
  offset = NULL,
  method = "Fisher",
  type = "forward",
  metric = "AIC",
  keep = NULL,
  maxsize = NULL,
  grads = 10,
  parallel = FALSE,
  nthreads = 8,
  tol = 1e-04,
  contrasts = NULL,
  showprogress = TRUE
)

\method{VariableSelection}{BranchGLM}(
  fit,
  type = "forward",
  metric = "AIC",
  keep = NULL,
  maxsize = NULL,
  method = NULL,
  grads = 10,
  parallel = FALSE,
  nthreads = 8,
  tol = 1e-04,
  showprogress = TRUE
)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A dataframe.}

\item{family}{distribution used to model the data, one of "gaussian", "binomal", or "poisson"}

\item{link}{link used to link mean structure to linear predictors. One of, 
"identity", "logit", "probit", "cloglog", or "log".}

\item{offset}{Offset vector, by default the zero vector is used.}

\item{method}{One of "Fisher", "BFGS", or "LBFGS". If method is not specified, 
the method that was originally used to fit the BranchGLM model is used.}

\item{type}{One of "forward", "backward", or "branch and bound" to indicate which type of variable selection to perform.}

\item{metric}{Metric used to choose model, the default is "AIC".}

\item{keep}{Vector of names to denote variables that must be in the model.}

\item{maxsize}{Maximum number of variables to consider in a single model, the 
default is the total number of variables.
This number adds onto any variables specified in keep.}

\item{grads}{Number of gradients to use to approximate information with, only for LBFGS.}

\item{parallel}{One of TRUE or FALSE to indicate if parallelization should be used.
Only available for branch and bound selection.}

\item{nthreads}{Number of threads used with OpenMP, only used if parallel is TRUE.}

\item{tol}{Tolerance used to determine model convergence.}

\item{showprogress}{Whether to show progress updates for branch and bound.}

\item{fit}{A BranchGLM object.}
}
\value{
A BranchGLMVS object.
}
\description{
Performs variable selection for GLMs with forward, backward, or 
branch and bound selection. Performs best subset selection using a branch and 
bound algorithm that can massively speed up the process.
}
\details{
The model in the formula or the formula from the fitted model is 
treated as the upper model. The variables specified in keep along with an 
intercept (if included in formula) is the lower model. When an intercept is 
included in the model formula it is kept in each model.
}
