% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{BranchGLM}
\alias{BranchGLM}
\title{Fits GLMs}
\usage{
BranchGLM(
  formula,
  data,
  family,
  link,
  offset = NULL,
  method = "Fisher",
  grads = 10,
  parallel = FALSE,
  nthreads = 8,
  tol = 1e-04,
  maxit = NULL,
  contrasts = NULL
)
}
\arguments{
\item{formula}{a formula for the model.}

\item{data}{a dataframe that contains the response and predictor variables.}

\item{family}{distribution used to model the data, one of "gaussian", "binomial", or "poisson".}

\item{link}{link used to link mean structure to linear predictors. One of, 
"identity", "logit", "probit", "cloglog", or "log".}

\item{offset}{offset vector, by default the zero vector is used.}

\item{method}{one of "Fisher", "BFGS", or "LBFGS".}

\item{grads}{number of gradients to use to approximate information with, only for LBFGS.}

\item{parallel}{whether or not to make use of parallelization via OpenMP.}

\item{nthreads}{number of threads used with OpenMP, only used if parallel = TRUE.}

\item{tol}{tolerance used to determine model convergence.}

\item{maxit}{maximum number of iterations performed. The default for 
Fisher's scoring is 50 and for the other methods the default is 200.}

\item{contrasts}{see \code{contrasts.arg} of \code{model.matrix.default}.}
}
\value{
A \code{BranchGLM} object which is a list with the following components
\item{\code{coefficients}}{ a matrix with the coefficients estimates, SEs, wald test statistics, and p-values}
\item{\code{iterations}}{ number of iterations it took the algorithm to converge, if the algorithm failed to converge then this is -1}
\item{\code{dispersion}}{ the value of the dispersion parameter}
\item{\code{logLik}}{ the log-likelihood of the fitted model}
\item{\code{resdev}}{ the residual deviance of the fitted model}
\item{\code{AIC}}{ the AIC of the fitted model}
\item{\code{preds}}{ predictions from the fitted model}
\item{\code{linpreds}}{ linear predictors from the fitted model}
\item{\code{formula}}{ formula used to fit the model}
\item{\code{method}}{ iterative method used to fit the model}
\item{\code{y}}{ y vector used in the model}
\item{\code{x}}{ design matrix used to fit the model}
\item{\code{data}}{ original dataframe supplied to the function}
\item{\code{names}}{ names of the variables}
\item{\code{yname}}{ name of y variable}
\item{\code{parallel}}{ whether parallelization was employed to speed up model fitting process}
\item{\code{missing}}{ number of missing values removed from the original dataset}
\item{\code{link}}{ link function used to model the data}
\item{\code{offset}}{ offset vector}
\item{\code{family}}{ family used to model the data}
\item{\code{ylevel}}{ the levels of y, only included for binomial glms}
}
\description{
Fits generalized linear models via 'RcppArmadillo'. Also has the 
ability to fit the models with parallelization via 'OpenMP'.
}
\details{
Can use BFGS, L-BFGS, or Fisher's scoring to fit the GLM. BFGS and L-BFGS are 
typically faster than Fisher's scoring when there are at least 50 covariates 
and Fisher's scoring is typically best when there are fewer than 20 covariates.
This function does not currently support the use of weights. 

The models are fit in C++ by using 'Rcpp' and 'RcppArmadillo'. In order to help 
convergence, each of the methods makes use of a backtracking line-search using 
the armijo-goldstein condition to find an adequate step size. There are also 
two conditions used to control convergence, the first is whether there is a 
sufficient decrease in the negative log-likelihood, and the other is whether 
each of the elements of the beta vector changes by a sufficient amount. The 
\code{tol} argument controls both of these criteria. If the algorithm fails to 
converge, then \code{iterations} will be -1.

The likelihood equations are solved directly, i.e. no matrix decomposition is used.
All observations with any missing values are ignored.
}
\examples{
Data <- iris
BranchGLM(Sepal.Length ~ ., data = Data, family = "gaussian", link = "identity")
}
