% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMTable.R
\name{Cindex}
\alias{Cindex}
\alias{AUC}
\alias{Cindex.numeric}
\alias{Cindex.BranchGLM}
\alias{Cindex.BranchGLMROC}
\title{Cindex/AUC}
\usage{
Cindex(x, ...)

AUC(x, ...)

\method{Cindex}{numeric}(preds, y)

\method{Cindex}{BranchGLM}(fit)

\method{Cindex}{BranchGLMROC}(ROC)
}
\arguments{
\item{preds}{A numeric vector of predictions between 0 and 1.}

\item{y}{Observed values, can be a numeric vector of 0s and 1s, a two-level 
factor vector, or a logical vector.}

\item{fit}{A BranchGLM object.}

\item{ROC}{A BranchGLMROC object.}
}
\value{
The c-index or AUC.
}
\description{
Calculates c-index.
}
\details{
Uses trapezoidal rule to calculate AUC when given a BranchGLMROC object.
Uses Mann-Whitney U to calculate it otherwise.
}
