% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{coef.BranchGLM}
\alias{coef.BranchGLM}
\title{Extract Coefficients}
\usage{
\method{coef}{BranchGLM}(object, ...)
}
\arguments{
\item{object}{a \code{BranchGLM} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A named vector with the corresponding coefficient estimates.
}
\description{
Extract Coefficients
}
