% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection.R
\name{coef.BranchGLMVS}
\alias{coef.BranchGLMVS}
\title{Extract Coefficients}
\usage{
\method{coef}{BranchGLMVS}(object, which = 1, ...)
}
\arguments{
\item{object}{a \code{BranchGLMVS} object.}

\item{which}{which models to get coefficients from, the default is the best model. 
Can specify "all" to get coefficients from all of the best models.}

\item{...}{further arguments to \link{fit.BranchGLMVS}.}
}
\value{
A numeric matrix with the corresponding coefficient estimates.
}
\description{
Extract Coefficients
}
