% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{formula.BranchGLM}
\alias{formula.BranchGLM}
\title{Extract Model Formula from BranchGLM Objects}
\usage{
\method{formula}{BranchGLM}(x, ...)
}
\arguments{
\item{x}{a \code{BranchGLM} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A formula representing the model used to obtain \code{x}.
}
\description{
Extracts model formula from \code{BranchGLM} objects.
}
