% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{logLik.BranchGLM}
\alias{logLik.BranchGLM}
\title{Extract Log-Likelihood from BranchGLM Objects}
\usage{
\method{logLik}{BranchGLM}(object, ...)
}
\arguments{
\item{object}{a \code{BranchGLM} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \code{logLik} which is a number corresponding to
the log-likelihood with the following attributes: "df" (degrees of freedom)
and "nobs" (number of observations).
}
\description{
Extracts log-likelihood from BranchGLM objects.
}
