% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMTable.R
\name{MultipleROCCurves}
\alias{MultipleROCCurves}
\title{Plotting Multiple ROC Curves}
\usage{
MultipleROCCurves(
  ...,
  legendpos = "bottomright",
  title = "ROC Curves",
  colors = NULL,
  names = NULL,
  lty = 1,
  lwd = 1
)
}
\arguments{
\item{...}{Any number of BranchGLMROC objects.}

\item{legendpos}{Where to place legend.}

\item{title}{Title for plot.}

\item{colors}{Vector of colors to be used on ROC curves.}

\item{names}{Vector of names used to create legend for ROC curves.}

\item{lty}{Vector of linetypes used to create legend for ROC curves or a 
single linetype to be used for all ROC curves.}

\item{lwd}{Vector of linewidths used to create legend for ROC curves or a 
single linewidth to be used for all ROC curves.}
}
\description{
Plotting Multiple ROC Curves
}
