% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMTable.R
\name{plot.BranchGLMROC}
\alias{plot.BranchGLMROC}
\title{Plotting ROC Curve}
\usage{
\method{plot}{BranchGLMROC}(x, ...)
}
\arguments{
\item{x}{a \code{BranchGLMROC} object.}

\item{...}{arguments passed to generic plot function.}
}
\value{
No return value, called to create the plot.
}
\description{
This plots a ROC curve.
}
\examples{
Data <- ToothGrowth
Fit <- BranchGLM(supp ~ ., data = Data, family = "binomial", link = "logit")
MyROC <- ROC(Fit)
plot(MyROC)
}
