% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMTable.R
\name{plot.BranchGLMROC}
\alias{plot.BranchGLMROC}
\title{Plotting ROC Curve}
\usage{
\method{plot}{BranchGLMROC}(ROC, ...)
}
\arguments{
\item{ROC}{A BranchGLMROC object.}

\item{...}{arguments passed to generic plot function.}
}
\description{
plot plots a ROC curve.
}
