% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection.R
\name{predict.BranchGLMVS}
\alias{predict.BranchGLMVS}
\title{Predict Method for BranchGLMVS Objects}
\usage{
\method{predict}{BranchGLMVS}(object, newdata = NULL, type = "response", which = 1, ...)
}
\arguments{
\item{object}{a \code{BranchGLMVS} object.}

\item{newdata}{a dataframe, if not specified then the data the model was fit on is used.}

\item{type}{one of "linpreds" or "response", if not specified then "response" is used.}

\item{which}{which model to get predictions from, the default is the best model.}

\item{...}{further arguments passed to \link{fit.BranchGLMVS}.}
}
\value{
A numeric vector of predictions.
}
\description{
Gets predictions from a \code{BranchGLMVS} object.
}
\details{
linpreds corresponds to the linear predictors and response is on the scale of the response variable.
Offset variables are ignored for predictions on new data.
}
