% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{print.BranchGLM}
\alias{print.BranchGLM}
\title{Print Method for BranchGLM}
\usage{
\method{print}{BranchGLM}(fit, coefdigits = 4, digits = 0)
}
\arguments{
\item{fit}{A BranchGLM model object.}

\item{coefdigits}{Number of digits to display for coefficients table.}

\item{digits}{Number of digits to display for information after table.}
}
\description{
Print Method for BranchGLM
}
