% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMCIs.R
\name{print.BranchGLMCIs}
\alias{print.BranchGLMCIs}
\title{Print Method for BranchGLMCIs Objects}
\usage{
\method{print}{BranchGLMCIs}(object, digits = 4, ...)
}
\arguments{
\item{object}{a \code{BranchGLMCIs} object.}

\item{digits}{number of significant digits to display.}

\item{...}{further arguments passed from other methods.}
}
\value{
The supplied \code{BranchGLMCIs} object.
}
\description{
Print Method for BranchGLMCIs Objects
}
