% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMTable.R
\name{print.BranchGLMROC}
\alias{print.BranchGLMROC}
\title{Print Method for BranchGLMROC}
\usage{
\method{print}{BranchGLMROC}(ROC)
}
\arguments{
\item{ROC}{BranchGLMROC object.}
}
\value{
a BranchGLMROC object.
}
\description{
Print Method for BranchGLMROC
}
