% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMTable.R
\name{print.BranchGLMROC}
\alias{print.BranchGLMROC}
\title{Print Method for BranchGLMROC Objects}
\usage{
\method{print}{BranchGLMROC}(x, ...)
}
\arguments{
\item{x}{a \code{BranchGLMROC} object.}

\item{...}{further arguments passed to other methods.}
}
\value{
The supplied \code{BranchGLMROC} object.
}
\description{
Print method for BranchGLMROC objects.
}
