% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMTable.R
\name{print.BranchGLMTable}
\alias{print.BranchGLMTable}
\title{Print Method for BranchGLMTable}
\usage{
\method{print}{BranchGLMTable}(x, digits = 4, ...)
}
\arguments{
\item{x}{a \code{BranchGLMTable} object.}

\item{digits}{number of digits to display.}

\item{...}{further arguments passed to other methods.}
}
\description{
Print Method for BranchGLMTable
}
