% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection.R
\name{print.BranchGLMVS}
\alias{print.BranchGLMVS}
\title{Print Method for BranchGLMVS}
\usage{
\method{print}{BranchGLMVS}(x, coefdigits = 4, digits = 0, ...)
}
\arguments{
\item{x}{a \code{BranchGLMVS} object.}

\item{coefdigits}{number of digits to display for coefficients table.}

\item{digits}{number of digits to display for information not in the table.}

\item{...}{further arguments passed to other methods.}
}
\description{
Print Method for BranchGLMVS
}
