% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLMTable.R
\name{Table}
\alias{Table}
\alias{Table.numeric}
\alias{Table.factor}
\alias{Table.BranchGLM}
\title{Confusion Matrix}
\usage{
Table(x, ...)

\method{Table}{numeric}(preds, y, cutoff = 0.5)

\method{Table}{factor}(preds, y)

\method{Table}{BranchGLM}(fit, cutoff = 0.5)
}
\arguments{
\item{preds}{a numeric vector of predictions between 0 and 1 or a two-level factor vector.}

\item{y}{observed values, can be a numeric vector of 0s and 1s, a two-level factor vector, or 
a logical vector. Must be a two-level factor vector if preds is also a two-level factor vector.}

\item{cutoff}{cutoff for predicted values, default is 0.5.}

\item{fit}{a \code{BranchGLM} object.}
}
\value{
a \code{BranchGLMTable} object.
}
\description{
Creates confusion matrix and calculates related measures.
}
\examples{
Data <- ToothGrowth
Fit <- BranchGLM(supp ~ ., data = Data, family = "binomial", link = "logit")
Table(Fit)
}
