% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchGLM.R
\name{vcov.BranchGLM}
\alias{vcov.BranchGLM}
\title{Extract covariance matrix}
\usage{
\method{vcov}{BranchGLM}(object, ...)
}
\arguments{
\item{object}{a \code{BranchGLM} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A numeric matrix which is the covariance matrix of the beta coefficients.
}
\description{
Extract covariance matrix
}
